remove intermediate ControlledHighTable
